+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [feature] regex tester add error message and more options.
 * [feature] add favorite collection.
 * [feature] add UUID v5.
+* [feature] blurhash to image.
 * [fix] fix password generator algorithm.
 * [fix] fix regex tester memory leak.